scripts for graph-codegen and graph-deploy'
