+++ package.json
@@ -12,9 +12,9 @@
 		"node": ">=10.13.0"
 	},
 	"scripts": {
-		"codegen": "node ops/generate-abis.js && node ops/generate-schema.js && node ops/generate-subgraph.js && graph codegen --output-dir src/types/ subgraph.yaml",
+		"codegen": "node ops/generate-abis.js && node ops/generate-schema.js && node ops/generate-subgraph.js && node ops/graph-codegen.js",
 		"codegen:watch": "watch 'npm run codegen' src/mappings src/domain",
-		"deploy": "npm run migrate && npm run codegen && graph deploy --access-token \"\" --ipfs ${ipfs-/ip4/127.0.0.1/tcp/5001} --node ${node_rpc-http://127.0.0.1:8020/} -n daostack subgraph.yaml",
+		"deploy": "npm run migrate && npm run codegen && node ops/graph-deploy",
 		"deploy:watch": "npm run deploy -- --watch",
 		"docker": "docker-compose run --rm -v $(pwd):/usr/app -v /usr/app/node_modules subgraph",
 		"docker:stop": "docker-compose down -v",
